finally delete EntityPropertyMapping
