+++ hibernate-core/src/main/java/org/hibernate/persister/entity/AbstractEntityPersister.java
@@ -229,6 +229,8 @@ import org.hibernate.type.CompositeType;
 import org.hibernate.type.EntityType;
 import org.hibernate.type.ManyToOneType;
 import org.hibernate.type.MappingContext;
+import org.hibernate.type.OneToOneType;
+import org.hibernate.type.SpecialOneToOneType;
 import org.hibernate.type.Type;
 import org.hibernate.type.descriptor.java.JavaType;
 import org.hibernate.type.spi.TypeConfiguration;
@@ -271,15 +273,18 @@ import static org.hibernate.generator.values.internal.GeneratedValuesHelper.getG
 import static org.hibernate.internal.CoreMessageLogger.CORE_LOGGER;
 import static org.hibernate.internal.util.ReflectHelper.isAbstractClass;
 import static org.hibernate.internal.util.StringHelper.isEmpty;
+import static org.hibernate.internal.util.StringHelper.qualifier;
 import static org.hibernate.internal.util.StringHelper.qualify;
 import static org.hibernate.internal.util.StringHelper.qualifyConditionally;
 import static org.hibernate.internal.util.StringHelper.replace;
 import static org.hibernate.internal.util.StringHelper.root;
 import static org.hibernate.internal.util.StringHelper.unqualify;
 import static org.hibernate.internal.util.collections.ArrayHelper.EMPTY_INT_ARRAY;
+import static org.hibernate.internal.util.collections.ArrayHelper.EMPTY_STRING_ARRAY;
 import static org.hibernate.internal.util.collections.ArrayHelper.contains;
 import static org.hibernate.internal.util.collections.ArrayHelper.indexOf;
 import static org.hibernate.internal.util.collections.ArrayHelper.isAllTrue;
+import static org.hibernate.internal.util.collections.ArrayHelper.slice;
 import static org.hibernate.internal.util.collections.ArrayHelper.to2DStringArray;
 import static org.hibernate.internal.util.collections.ArrayHelper.toIntArray;
 import static org.hibernate.internal.util.collections.ArrayHelper.toStringArray;
@@ -461,12 +466,12 @@ public abstract class AbstractEntityPersister
 	/**
 	 * Warning:
 	 * When there are duplicated property names in the subclasses
-	 * then propertyMapping will only contain one of those properties.
-	 * To ensure correct results, propertyMapping should only be used
+	 * then propertyPathRegistry will only contain one of those properties.
+	 * To ensure correct results, propertyPathRegistry should only be used
 	 * for the concrete EntityPersister (since the concrete EntityPersister
 	 * cannot have duplicated property names).
 	 */
-	private final EntityPropertyMapping propertyMapping;
+	private final PropertyPathRegistry propertyPathRegistry;
 
 	private List<UniqueKeyEntry> uniqueKeyEntries = null; //lazily initialized
 	private ConcurrentHashMap<String,SingleIdArrayLoadPlan> nonLazyPropertyLoadPlansByName;
@@ -559,7 +564,7 @@ public abstract class AbstractEntityPersister
 		hasPartitionedSelectionMapping = persistentClass.hasPartitionedSelectionMapping();
 		hasCollectionNotReferencingPK = persistentClass.hasCollectionNotReferencingPK();
 
-		propertyMapping = new EntityPropertyMapping( this );
+		propertyPathRegistry = new PropertyPathRegistry();
 
 		// IDENTIFIER
 
@@ -2381,7 +2386,7 @@ public abstract class AbstractEntityPersister
 	 */
 	@Override
 	public String[] toColumns(String propertyName) throws QueryException {
-		return propertyMapping.getColumnNames( propertyName );
+		return propertyPathRegistry.getColumnNames( propertyName );
 	}
 
 	/**
@@ -2396,7 +2401,7 @@ public abstract class AbstractEntityPersister
 	 */
 	@Override
 	public String[] getPropertyColumnNames(String propertyName) {
-		return propertyMapping.getColumnNames( propertyName );
+		return propertyPathRegistry.getColumnNames( propertyName );
 	}
 
 	private DiscriminatorType<?> discriminatorDomainType;
@@ -2706,7 +2711,7 @@ public abstract class AbstractEntityPersister
 
 	private void initOrdinaryPropertyPaths(Metadata mapping) throws MappingException {
 		for ( int i = 0; i < getSubclassPropertyNameClosure().length; i++ ) {
-			propertyMapping.initPropertyPaths(
+			propertyPathRegistry.initPropertyPaths(
 					getSubclassPropertyNameClosure()[i],
 					getSubclassPropertyTypeClosure()[i],
 					getSubclassPropertyColumnNameClosure()[i],
@@ -2721,19 +2726,19 @@ public abstract class AbstractEntityPersister
 	private void initIdentifierPropertyPaths(Metadata mapping) throws MappingException {
 		final String idProp = getIdentifierPropertyName();
 		if ( idProp != null ) {
-			propertyMapping.initPropertyPaths(
+			propertyPathRegistry.initPropertyPaths(
 					idProp, getIdentifierType(), getIdentifierColumnNames(),
 					getIdentifierColumnReaders(), getIdentifierColumnReaderTemplates(), null, mapping
 			);
 		}
 		if ( isIdentifierEmbedded() ) {
-			propertyMapping.initPropertyPaths(
+			propertyPathRegistry.initPropertyPaths(
 					null, getIdentifierType(), getIdentifierColumnNames(),
 					getIdentifierColumnReaders(), getIdentifierColumnReaderTemplates(), null, mapping
 			);
 		}
 		if ( !hasNonIdentifierPropertyNamedId() ) {
-			propertyMapping.initPropertyPaths(
+			propertyPathRegistry.initPropertyPaths(
 					ENTITY_ID, getIdentifierType(), getIdentifierColumnNames(),
 					getIdentifierColumnReaders(), getIdentifierColumnReaderTemplates(), null, mapping
 			);
@@ -2741,7 +2746,7 @@ public abstract class AbstractEntityPersister
 	}
 
 	private void initDiscriminatorPropertyPath(Metadata mapping) {
-		propertyMapping.initPropertyPaths(
+		propertyPathRegistry.initPropertyPaths(
 				ENTITY_CLASS,
 				getDiscriminatorType(),
 				new String[] {getDiscriminatorColumnName()},
@@ -2758,6 +2763,322 @@ public abstract class AbstractEntityPersister
 		initIdentifierPropertyPaths( mapping );
 		if ( isPolymorphic() ) {
 			initDiscriminatorPropertyPath( mapping );
+		}
+	}
+
+	private final class PropertyPathRegistry {
+		private final Map<String, Type> typesByPropertyPath = new HashMap<>();
+		private final Map<String, String[]> columnsByPropertyPath = new HashMap<>();
+		private final Map<String, String[]> columnReadersByPropertyPath = new HashMap<>();
+		private final Map<String, String[]> columnReaderTemplatesByPropertyPath = new HashMap<>();
+
+		// This field is only used during initialization, no need for threadsafety.
+		private Set<String> duplicateIncompatiblePaths;
+
+		Type toType(String propertyName) throws QueryException {
+			Type type = typesByPropertyPath.get( propertyName );
+			if ( type == null ) {
+				throw propertyException( propertyName );
+			}
+			return type;
+		}
+
+		private QueryException propertyException(String propertyName) throws QueryException {
+			return new QueryException( "Could not resolve property: " + propertyName + " of: " + getEntityName() );
+		}
+
+		String[] getColumnNames(String propertyName) {
+			final var cols = columnsByPropertyPath.get( propertyName );
+			if ( cols == null ) {
+				throw new MappingException( "Unknown property: " + propertyName );
+			}
+			return cols;
+		}
+
+		private void logIncompatibleRegistration(String path, Type existingType, Type type) {
+			if ( CORE_LOGGER.isTraceEnabled() ) {
+				CORE_LOGGER.tracev(
+						"Skipped adding attribute [{1}] to base type [{0}] as more than one subtype defined the attribute using incompatible types (strictly speaking the attributes are not inherited); existing type = [{2}], incoming type = [{3}]",
+						getEntityName(),
+						path,
+						existingType,
+						type
+				);
+			}
+		}
+
+		private void addPropertyPath(
+				String path,
+				Type type,
+				String[] columns,
+				String[] columnReaders,
+				String[] columnReaderTemplates,
+				Metadata mapping) {
+			final Type existingType = typesByPropertyPath.get( path );
+			if ( existingType != null
+					|| ( duplicateIncompatiblePaths != null && duplicateIncompatiblePaths.contains( path ) ) ) {
+				// If types match or the new type is not an association type, there is nothing for us to do.
+				if ( type == existingType || existingType == null || !( type instanceof AssociationType ) ) {
+				}
+				else if ( !( existingType instanceof AssociationType ) ) {
+					// Workaround for org.hibernate.cfg.annotations.PropertyBinder#bind() adding a component for *ToOne ids.
+				}
+				else {
+					if ( type instanceof AnyType && existingType instanceof AnyType ) {
+						// Let the first type dictate what type the property has.
+					}
+					else {
+						Type commonType = null;
+						final var metadata = (MetadataImplementor) mapping;
+						if ( type instanceof CollectionType collectionType
+								&& existingType instanceof CollectionType existingCollectionType ) {
+							final var thisCollection = metadata.getCollectionBinding( existingCollectionType.getRole() );
+							final var otherCollection = metadata.getCollectionBinding( collectionType.getRole() );
+							if ( thisCollection.isSame( otherCollection ) ) {
+								return;
+							}
+							else {
+								logIncompatibleRegistration( path, existingType, type );
+							}
+						}
+						else if ( type instanceof EntityType entityType2 && existingType instanceof EntityType entityType1 ) {
+							if ( entityType1.getAssociatedEntityName().equals( entityType2.getAssociatedEntityName() ) ) {
+								return;
+							}
+							else {
+								commonType = getCommonType( metadata, entityType1, entityType2 );
+							}
+						}
+						else {
+							logIncompatibleRegistration( path, existingType, type );
+						}
+						if ( commonType == null ) {
+							if ( duplicateIncompatiblePaths == null ) {
+								duplicateIncompatiblePaths = new HashSet<>();
+							}
+							duplicateIncompatiblePaths.add( path );
+							typesByPropertyPath.remove( path );
+							columnsByPropertyPath.put( path, EMPTY_STRING_ARRAY );
+							columnReadersByPropertyPath.put( path, EMPTY_STRING_ARRAY );
+							columnReaderTemplatesByPropertyPath.put( path, EMPTY_STRING_ARRAY );
+						}
+						else {
+							typesByPropertyPath.put( path, commonType );
+						}
+					}
+				}
+			}
+			else {
+				typesByPropertyPath.put( path, type );
+				columnsByPropertyPath.put( path, columns );
+				columnReadersByPropertyPath.put( path, columnReaders );
+				columnReaderTemplatesByPropertyPath.put( path, columnReaderTemplates );
+			}
+		}
+
+		private Type getCommonType(MetadataImplementor metadata, EntityType entityType1, EntityType entityType2) {
+			final var thisClass = metadata.getEntityBinding( entityType1.getAssociatedEntityName() );
+			final var otherClass = metadata.getEntityBinding( entityType2.getAssociatedEntityName() );
+			final var commonClass = getCommonPersistentClass( thisClass, otherClass );
+
+			if ( commonClass == null ) {
+				return null;
+			}
+			else if ( entityType1 instanceof ManyToOneType manyToOneType ) {
+				return new ManyToOneType( manyToOneType, commonClass.getEntityName() );
+			}
+			else if ( entityType1 instanceof SpecialOneToOneType specialOneToOneType ) {
+				return new SpecialOneToOneType( specialOneToOneType, commonClass.getEntityName() );
+			}
+			else if ( entityType1 instanceof OneToOneType oneToOneType ) {
+				return new OneToOneType( oneToOneType, commonClass.getEntityName() );
+			}
+			else {
+				throw new IllegalStateException( "Unexpected entity type: " + entityType1 );
+			}
+		}
+
+		private PersistentClass getCommonPersistentClass(PersistentClass clazz1, PersistentClass clazz2) {
+			while ( clazz2 != null
+					&& clazz2.getMappedClass() != null
+					&& clazz1.getMappedClass() != null
+					&& !clazz2.getMappedClass().isAssignableFrom( clazz1.getMappedClass() ) ) {
+				clazz2 = clazz2.getSuperclass();
+			}
+			return clazz2;
+		}
+
+		private void initPropertyPaths(
+				final String path,
+				final Type type,
+				String[] columns,
+				String[] columnReaders,
+				String[] columnReaderTemplates,
+				final String[] formulaTemplates,
+				final Metadata mapping) throws MappingException {
+			assert columns != null : "Incoming columns should not be null : " + path;
+			assert type != null : "Incoming type should not be null : " + path;
+
+			if ( columns.length != type.getColumnSpan( mapping ) ) {
+				throw new MappingException(
+						"broken column mapping for: " + path +
+								" of: " + getEntityName()
+				);
+			}
+
+			if ( type instanceof AnyType || type instanceof CollectionType || type instanceof EntityType ) {
+				final var associationType = (AssociationType) type;
+				if ( associationType.useLHSPrimaryKey() ) {
+					columns = getIdentifierColumnNames();
+					columnReaders = getIdentifierColumnReaders();
+					columnReaderTemplates = getIdentifierColumnReaderTemplates();
+				}
+				else {
+					final String foreignKeyProperty = associationType.getLHSPropertyName();
+					if ( foreignKeyProperty != null && !path.equals( foreignKeyProperty ) ) {
+						// This requires the referenced property to have been processed first.
+						columns = columnsByPropertyPath.get( foreignKeyProperty );
+						if ( columns == null ) {
+							return;
+						}
+						columnReaders = columnReadersByPropertyPath.get( foreignKeyProperty );
+						columnReaderTemplates = columnReaderTemplatesByPropertyPath.get( foreignKeyProperty );
+					}
+				}
+			}
+
+			if ( path != null ) {
+				addPropertyPath( path, type, columns, columnReaders, columnReaderTemplates, mapping );
+			}
+
+			if ( type instanceof AnyType anyType ) {
+				initComponentPropertyPaths(
+						path,
+						anyType,
+						columns,
+						columnReaders,
+						columnReaderTemplates,
+						formulaTemplates,
+						mapping
+				);
+			}
+			else if ( type instanceof ComponentType componentType ) {
+				initComponentPropertyPaths(
+						path,
+						componentType,
+						columns,
+						columnReaders,
+						columnReaderTemplates,
+						formulaTemplates,
+						mapping
+				);
+				if ( componentType.isEmbedded() ) {
+					initComponentPropertyPaths(
+							path == null ? null : qualifier( path ),
+							componentType,
+							columns,
+							columnReaders,
+							columnReaderTemplates,
+							formulaTemplates,
+							mapping
+					);
+				}
+			}
+			else if ( type instanceof EntityType entityType ) {
+				initIdentifierPropertyPaths(
+						path,
+						entityType,
+						columns,
+						columnReaders,
+						columnReaderTemplates,
+						formulaTemplates != null && formulaTemplates.length > 0 ? formulaTemplates : null,
+						mapping
+				);
+			}
+		}
+
+		private void initIdentifierPropertyPaths(
+				final String path,
+				final EntityType entityType,
+				final String[] columns,
+				final String[] columnReaders,
+				final String[] columnReaderTemplates,
+				final String[] formulaTemplates,
+				final Metadata mapping) throws MappingException {
+
+			final var idtype = entityType.getIdentifierOrUniqueKeyType( mapping );
+			final String idPropName = entityType.getIdentifierOrUniqueKeyPropertyName( mapping );
+			boolean hasNonIdentifierPropertyNamedId = hasNonIdentifierPropertyNamedId( entityType, mapping );
+
+			if ( entityType.isReferenceToPrimaryKey() && !hasNonIdentifierPropertyNamedId ) {
+				final String idpath = extendPath( path, ENTITY_ID );
+				addPropertyPath( idpath, idtype, columns, columnReaders, columnReaderTemplates, mapping );
+				initPropertyPaths( idpath, idtype, columns, columnReaders, columnReaderTemplates, formulaTemplates,
+						mapping );
+			}
+
+			if ( !entityType.isNullable() && idPropName != null ) {
+				final String idpath2 = extendPath( path, idPropName );
+				addPropertyPath( idpath2, idtype, columns, columnReaders, columnReaderTemplates, mapping );
+				initPropertyPaths( idpath2, idtype, columns, columnReaders, columnReaderTemplates, formulaTemplates,
+						mapping );
+			}
+		}
+
+		private boolean hasNonIdentifierPropertyNamedId(final EntityType entityType, final MappingContext mapping) {
+			try {
+				return mapping.getReferencedPropertyType(
+						entityType.getAssociatedEntityName(),
+						ENTITY_ID
+				) != null;
+			}
+			catch (MappingException e) {
+				return false;
+			}
+		}
+
+		private void initComponentPropertyPaths(
+				final String path,
+				final CompositeType type,
+				final String[] columns,
+				final String[] columnReaders,
+				final String[] columnReaderTemplates,
+				final String[] formulaTemplates,
+				final Metadata mapping) throws MappingException {
+
+			final var types = type.getSubtypes();
+			final var properties = type.getPropertyNames();
+			int begin = 0;
+			for ( int i = 0; i < properties.length; i++ ) {
+				final String subpath = extendPath( path, properties[i] );
+				try {
+					final int length = types[i].getColumnSpan( mapping );
+					final var columnSlice = slice( columns, begin, length );
+					final var columnReaderSlice = slice( columnReaders, begin, length );
+					final var columnReaderTemplateSlice = slice( columnReaderTemplates, begin, length );
+					final var formulaSlice =
+							formulaTemplates == null
+									? null
+									: slice( formulaTemplates, begin, length );
+					initPropertyPaths(
+							subpath,
+							types[i],
+							columnSlice,
+							columnReaderSlice,
+							columnReaderTemplateSlice,
+							formulaSlice,
+							mapping
+					);
+					begin += length;
+				}
+				catch (Exception e) {
+					throw new MappingException( "bug in initComponentPropertyPaths", e );
+				}
+			}
+		}
+
+		private String extendPath(String path, String property) {
+			return isEmpty( path ) ? property : qualify( path, property );
 		}
 	}
 
@@ -4531,7 +4852,7 @@ public abstract class AbstractEntityPersister
 	@Override @Deprecated
 	public Type getPropertyType(String propertyName) throws MappingException {
 		// todo (PropertyMapping) : caller also deprecated (aka, easy to remove)
-		return propertyMapping.toType( propertyName );
+		return propertyPathRegistry.toType( propertyName );
 	}
 
 	@Override
